+++ core/src/com/color/game/utils/BodyUtils.java
@@ -28,6 +28,11 @@ public class BodyUtils {
         return userData != null && (userData.getUserDataType() == UserDataType.COLORPLATFORM || userData.getUserDataType() == UserDataType.PLATFORM);
     }
 
+    public static boolean isEnemy(Body body){
+        UserData userData = (UserData) body.getUserData();
+        return userData != null && userData.getUserDataType() == UserDataType.ENEMY;
+    }
+
     public static boolean onWall(Body character, Body platform){
         boolean onWall = false;